fix: Move to pages/api
