--- apis/info/.dev.vars
@@ -1,2 +1,0 @@
-BSC_NODE=https://bsc-dataseed.binance.org
-BIT_QUERY_HEADER=
