--- apis/info/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "rules": {
-    "no-restricted-globals": "off"
-  }
-}
