--- apis/info/src/bindings.d.ts
@@ -1,6 +1,0 @@
-export {}
-
-declare global {
-  const BSC_NODE: string
-  const BIT_QUERY_HEADER: string
-}
