--- apis/info/src/getBlocksFromTimestamps.ts
@@ -1,58 +1,0 @@
-import orderBy from 'lodash/orderBy'
-import { gql } from 'graphql-request'
-import { multiQuery } from './infoQueryHelpers'
-
-const BSC_BLOCK_CLIENT = 'https://api.thegraph.com/subgraphs/name/pancakeswap/blocks'
-
-const blocksQueryConstructor = (subqueries: string[]) => {
-  return gql`query blocks {
-    ${subqueries}
-  }`
-}
-
-const getBlockSubqueries = (timestamps: number[]) =>
-  timestamps.map((timestamp) => {
-    return `t${timestamp}:blocks(first: 1, orderBy: timestamp, orderDirection: desc, where: { timestamp_gt: ${timestamp}, timestamp_lt: ${
-      timestamp + 600
-    } }) {
-      number
-    }`
-  })
-
-interface Block {
-  number: number
-  timestamp: string
-}
-
-export const getBlocksFromTimestamps = async (
-  timestamps: number[],
-  sortDirection: 'asc' | 'desc' | undefined = 'desc',
-  skipCount: number | undefined = 500,
-): Promise<Block[]> => {
-  if (timestamps?.length === 0) {
-    return []
-  }
-
-  const fetchedData: any = await multiQuery(
-    blocksQueryConstructor,
-    getBlockSubqueries(timestamps),
-    BSC_BLOCK_CLIENT,
-    skipCount,
-  )
-
-  const blocks: Block[] = []
-  if (fetchedData) {
-    // eslint-disable-next-line no-restricted-syntax
-    for (const key of Object.keys(fetchedData)) {
-      if (fetchedData[key].length > 0) {
-        blocks.push({
-          timestamp: key.split('t')[1],
-          number: parseInt(fetchedData[key][0].number, 10),
-        })
-      }
-    }
-    // graphql-request does not guarantee same ordering of batched requests subqueries, hence manual sorting
-    return orderBy(blocks, (block) => block.number, sortDirection)
-  }
-  return blocks
-}
