--- apis/info/src/infoQueryHelpers.ts
@@ -1,43 +1,0 @@
-import { GraphQLClient } from 'graphql-request'
-import requestWithTimeout from './requestWithTimeout'
-
-/**
- * Helper function to get large amount GraphQL subqueries
- * @param queryConstructor constructor function that combines subqueries
- * @param subqueries individual queries
- * @param endpoint GraphQL endpoint
- * @param skipCount how many subqueries to fire at a time
- * @returns
- */
-export const multiQuery = async (
-  queryConstructor: (subqueries: string[]) => string,
-  subqueries: string[],
-  endpoint: string | GraphQLClient,
-  skipCount = 1000,
-) => {
-  let fetchedData = {}
-  let allFound = false
-  let skip = 0
-  const client = typeof endpoint === 'string' ? new GraphQLClient(endpoint) : endpoint
-  try {
-    while (!allFound) {
-      let end = subqueries.length
-      if (skip + skipCount < subqueries.length) {
-        end = skip + skipCount
-      }
-      const subqueriesSlice = subqueries.slice(skip, end)
-      // eslint-disable-next-line no-await-in-loop
-      const result: any = await requestWithTimeout(client, queryConstructor(subqueriesSlice))
-      fetchedData = {
-        ...fetchedData,
-        ...result,
-      }
-      allFound = Object.keys(result).length < skipCount || skip + skipCount > subqueries.length
-      skip += skipCount
-    }
-    return fetchedData
-  } catch (error) {
-    console.error('Failed to fetch info data', error)
-    return null
-  }
-}
