--- apis/info/wrangler.toml
@@ -1,28 +1,0 @@
-compatibility_date = "2022-05-20"
-kv_namespaces = [
-  {binding = "INFO", id = "d8c9b05a13174eb59574425beceb2574", preview_id = "4ac09dba355c44fa858a67a22a9a63f1"},
-]
-main = "src/index.ts"
-name = "info-dev"
-node_compat = true
-[vars]
-FORCE_UPDATE_KEY = "force"
-KV_CACHE = false
-[triggers]
-crons = []
-
-[env.production]
-kv_namespaces = [
-  {binding = "INFO", id = "e67e7b9d27784f64a1c65af5e8eb1509"},
-]
-name = "info"
-node_compat = true
-[env.production.vars]
-KV_CACHE = true
-[env.production.triggers]
-crons = ["0 0 * * *", "*/1 * * * *"]
-
-# The necessary secrets are:
-# - BSC_NODE
-# - BIT_QUERY_HEADER
-# Run `echo <VALUE> | wrangler secret put <NAME>` for each of these
