+++ apps/web/src/pages/index.tsx
@@ -9,7 +9,7 @@ const IndexPage = () => {
 export const getStaticProps: GetStaticProps = async () => {
   const queryClient = new QueryClient()
   try {
-    const results = await (await fetch('http://127.0.0.1:8787/info')).json()
+    const results = await (await fetch('api/info')).json()
     queryClient.setQueryData(['totalTx30Days'], results.totalTx30Days)
     queryClient.setQueryData(['tvl'], results.tvl)
     queryClient.setQueryData(['addressCount30Days'], results.addressCount30Days)
